Build-less tooltips js component
