+++ src/js/index.js
@@ -6,4 +6,3 @@ import "./components/config";
 /* -----  Components -------------------------------------------------------- */
 
 import "./components/menu-search";
-import "./components/tooltip";
