--- www/assets/js/bundle-tooltip.js
@@ -1,2 +1,0 @@
-webpackJsonp([1],{DuR2:function(t,e){var r;r=function(){return this}();try{r=r||Function("return this")()||(0,eval)("this")}catch(t){"object"==typeof window&&(r=window)}t.exports=r},IDmg:function(t,e,r){"use strict";Object.defineProperty(e,"__esModule",{value:!0});var n=r("Zgw8");function i(){return(i=Object.assign||function(t){for(var e=1;e<arguments.length;e++){var r=arguments[e];for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&(t[n]=r[n])}return t}).apply(this,arguments)}var o="undefined"!=typeof window&&"undefined"!=typeof document,a=o?navigator.userAgent:"",p=/MSIE |Trident\//.test(a),s=/UCBrowser\//.test(a),l=o&&/iPhone|iPad|iPod/.test(navigator.platform)&&!window.MSStream,c={a11y:!0,allowHTML:!0,animateFill:!0,animation:"shift-away",appendTo:function(){return document.body},aria:"describedby",arrow:!1,arrowType:"sharp",boundary:"scrollParent",content:"",delay:0,distance:10,duration:[325,275],flip:!0,flipBehavior:"flip",flipOnUpdate:!1,followCursor:!1,hideOnClick:!0,ignoreAttributes:!1,inertia:!1,interactive:!1,interactiveBorder:2,interactiveDebounce:0,lazy:!0,maxWidth:350,multiple:!1,offset:0,onHidden:function(){},onHide:function(){},onMount:function(){},onShow:function(){},onShown:function(){},placement:"top",popperOptions:{},role:"tooltip",showOnInit:!1,size:"regular",sticky:!1,target:"",theme:"dark",touch:!0,touchHold:!1,trigger:"mouseenter focus",updateDuration:0,wait:null,zIndex:9999},f=["arrow","arrowType","boundary","distance","flip","flipBehavior","flipOnUpdate","offset","placement","popperOptions"],d=o?Element.prototype:{},u=d.matches||d.matchesSelector||d.webkitMatchesSelector||d.mozMatchesSelector||d.msMatchesSelector;function m(t){return[].slice.call(t)}function h(t,e){return(d.closest||function(t){for(var e=this;e;){if(u.call(e,t))return e;e=e.parentElement}}).call(t,e)}function b(t,e){for(;t;){if(e(t))return t;t=t.parentElement}}var v={passive:!0},y=4,g="x-placement",w="x-out-of-boundaries",x="tippy-iOS",k="tippy-active",E=".tippy-popper",O=".tippy-tooltip",L=".tippy-content",C=".tippy-backdrop",A=".tippy-arrow",T=".tippy-roundarrow",M=Object.keys(c);function Y(t,e){return{}.hasOwnProperty.call(t,e)}function I(t,e,r){if(Array.isArray(t)){var n=t[e];return null==n?r:n}return t}function S(t,e){var r;return function(){var n=this,i=arguments;clearTimeout(r),r=setTimeout(function(){return t.apply(n,i)},e)}}function D(t,e){return t&&t.modifiers&&t.modifiers[e]}function X(t,e){return t.indexOf(e)>-1}function H(t){return!(!t||!Y(t,"isVirtual"))||t instanceof Element}function N(t,e){return"function"==typeof t?t.apply(null,e):t}function F(t,e){t.filter(function(t){return"flip"===t.name})[0].enabled=e}function P(){return document.createElement("div")}function W(t,e){t.forEach(function(t){t&&(t.style.transitionDuration="".concat(e,"ms"))})}function B(t,e){t.forEach(function(t){t&&t.setAttribute("data-state",e)})}function z(t,e){var r=i({},e,{content:N(e.content,[t])},e.ignoreAttributes?{}:function(t){return M.reduce(function(e,r){var n=(t.getAttribute("data-tippy-".concat(r))||"").trim();if(!n)return e;if("content"===r)e[r]=n;else try{e[r]=JSON.parse(n)}catch(t){e[r]=n}return e},{})}(t));return(r.arrow||s)&&(r.animateFill=!1),r}function _(t,e){Object.keys(t).forEach(function(t){if(!Y(e,t))throw new Error("[tippy]: `".concat(t,"` is not a valid option"))})}function U(t,e){t.innerHTML=e instanceof Element?e.innerHTML:e}function j(t,e){if(e.content instanceof Element)U(t,""),t.appendChild(e.content);else if("function"!=typeof e.content){t[e.allowHTML?"innerHTML":"textContent"]=e.content}}function V(t){return{tooltip:t.querySelector(O),backdrop:t.querySelector(C),content:t.querySelector(L),arrow:t.querySelector(A)||t.querySelector(T)}}function R(t){t.setAttribute("data-inertia","")}function q(t){var e=P();return"round"===t?(e.className="tippy-roundarrow",U(e,'<svg viewBox="0 0 18 7" xmlns="http://www.w3.org/2000/svg"><path d="M0 7s2.021-.015 5.253-4.218C6.584 1.051 7.797.007 9 0c1.203-.007 2.416 1.035 3.761 2.782C16.012 7.005 18 7 18 7H0z"/></svg>')):e.className="tippy-arrow",e}function K(){var t=P();return t.className="tippy-backdrop",t.setAttribute("data-state","hidden"),t}function G(t,e){t.setAttribute("tabindex","-1"),e.setAttribute("data-interactive","")}function J(t,e,r){var n=s&&void 0!==document.body.style.webkitTransition?"webkitTransitionEnd":"transitionend";t[e+"EventListener"](n,r)}function Z(t){var e=t.getAttribute(g);return e?e.split("-")[0]:""}function Q(t,e,r){r.split(" ").forEach(function(r){t.classList[e](r+"-theme")})}function $(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},e=t.checkHideOnClick,r=t.exclude,n=t.duration;m(document.querySelectorAll(E)).forEach(function(t){var i=t._tippy;!i||e&&!0!==i.props.hideOnClick||r&&t===r.popper||i.hide(n)})}var tt=!1;function et(){tt||(tt=!0,l&&document.body.classList.add(x),window.performance&&document.addEventListener("mousemove",nt))}var rt=0;function nt(){var t=performance.now();t-rt<20&&(tt=!1,document.removeEventListener("mousemove",nt),l||document.body.classList.remove(x)),rt=t}function it(t){if(!(t.target instanceof Element))return $();var e=h(t.target,E);if(!(e&&e._tippy&&e._tippy.props.interactive)){var r=b(t.target,function(t){return t._tippy&&t._tippy.reference===t});if(r){var n=r._tippy;if(n){var i=X(n.props.trigger||"","click");if(tt||i)return $({exclude:n,checkHideOnClick:!0});if(!0!==n.props.hideOnClick||i)return;n.clearDelayTimeouts()}}$({checkHideOnClick:!0})}}function ot(){var t=document.activeElement;t&&t.blur&&t._tippy&&t.blur()}var at=1;function pt(t,e){var r,o,a,s,l,d=z(t,e);if(!d.multiple&&t._tippy)return null;var x,O,L,C,A=!1,T=!1,M=[],H=d.interactiveDebounce>0?S(mt,d.interactiveDebounce):mt,U=at++,$=function(t,e){var r=P();r.className="tippy-popper",r.id="tippy-".concat(t),r.style.zIndex=""+e.zIndex,e.role&&r.setAttribute("role",e.role);var n=P();n.className="tippy-tooltip",n.style.maxWidth=e.maxWidth+("number"==typeof e.maxWidth?"px":""),n.setAttribute("data-size",e.size),n.setAttribute("data-animation",e.animation),n.setAttribute("data-state","hidden"),Q(n,"add",e.theme);var i=P();return i.className="tippy-content",i.setAttribute("data-state","hidden"),e.interactive&&G(r,n),e.arrow&&n.appendChild(q(e.arrowType)),e.animateFill&&(n.appendChild(K()),n.setAttribute("data-animatefill","")),e.inertia&&R(n),j(i,e),n.appendChild(i),r.appendChild(n),r}(U,d),et={id:U,reference:t,popper:$,popperChildren:V($),popperInstance:null,props:d,state:{isEnabled:!0,isVisible:!1,isDestroyed:!1,isMounted:!1,isShown:!1},clearDelayTimeouts:Et,set:Ot,setContent:function(t){Ot({content:t})},show:Lt,hide:Ct,enable:function(){et.state.isEnabled=!0},disable:function(){et.state.isEnabled=!1},destroy:function(t){if(et.state.isDestroyed)return;et.state.isMounted&&Ct(0);ft(),delete et.reference._tippy,et.props.target&&t&&m(et.reference.querySelectorAll(et.props.target)).forEach(function(t){t._tippy&&t._tippy.destroy()});et.popperInstance&&et.popperInstance.destroy();et.state.isDestroyed=!0}};return t._tippy=et,$._tippy=et,ct(),d.lazy||(wt(),et.popperInstance.disableEventListeners()),d.showOnInit&&xt(),d.a11y&&!d.target&&((C=t)instanceof Element&&(!u.call(C,"a[href],area[href],button,details,input,textarea,select,iframe,[tabindex]")||C.hasAttribute("disabled")))&&t.setAttribute("tabindex","0"),$.addEventListener("mouseenter",function(t){et.props.interactive&&et.state.isVisible&&"mouseenter"===r&&xt(t)}),$.addEventListener("mouseleave",function(){et.props.interactive&&"mouseenter"===r&&document.addEventListener("mousemove",H)}),et;function rt(){document.removeEventListener("mousemove",dt)}function nt(){document.body.removeEventListener("mouseleave",kt),document.removeEventListener("mousemove",H)}function it(){return[et.popperChildren.tooltip,et.popperChildren.backdrop,et.popperChildren.content]}function ot(){return et.props.followCursor&&!tt&&"focus"!==r}function st(t,e){var r=et.popperChildren.tooltip;function n(t){t.target===r&&(J(r,"remove",n),e())}if(0===t)return e();J(r,"remove",L),J(r,"add",n),L=n}function lt(t,e){var r=arguments.length>2&&void 0!==arguments[2]&&arguments[2];et.reference.addEventListener(t,e,r),M.push({eventType:t,handler:e,options:r})}function ct(){et.props.touchHold&&!et.props.target&&(lt("touchstart",ut,v),lt("touchend",ht,v)),et.props.trigger.trim().split(" ").forEach(function(t){if("manual"!==t)if(et.props.target)switch(t){case"mouseenter":lt("mouseover",vt),lt("mouseout",yt);break;case"focus":lt("focusin",vt),lt("focusout",yt);break;case"click":lt(t,vt)}else switch(lt(t,ut),t){case"mouseenter":lt("mouseleave",ht);break;case"focus":lt(p?"focusout":"blur",bt)}})}function ft(){M.forEach(function(t){var e=t.eventType,r=t.handler,n=t.options;et.reference.removeEventListener(e,r,n)}),M=[]}function dt(t){var e=o=t,r=e.clientX,n=e.clientY;if(et.popperInstance){var a=Z(et.popper),p=et.props.arrow?y+("round"===et.props.arrowType?18:16):y,s=X(["top","bottom"],a),l=X(["left","right"],a),c=s?Math.max(p,r):r,f=l?Math.max(p,n):n;s&&c>p&&(c=Math.min(r,window.innerWidth-p)),l&&f>p&&(f=Math.min(n,window.innerHeight-p));var d=et.reference.getBoundingClientRect(),u=et.props.followCursor,m="horizontal"===u,h="vertical"===u;et.popperInstance.reference=i({},et.popperInstance.reference,{getBoundingClientRect:function(){return{width:0,height:0,top:m?d.top:f,bottom:m?d.bottom:f,left:h?d.left:c,right:h?d.right:c}},clientWidth:0,clientHeight:0}),et.popperInstance.scheduleUpdate(),"initial"===u&&et.state.isVisible&&rt()}}function ut(t){et.state.isEnabled&&!gt(t)&&(et.state.isVisible||(r=t.type,t instanceof MouseEvent&&(o=t)),"click"===t.type&&!1!==et.props.hideOnClick&&et.state.isVisible?kt():xt(t))}function mt(t){var e=b(t.target,function(t){return t._tippy}),r=h(t.target,E)===et.popper,n=e===et.reference;r||n||function(t,e,r,n){if(!t)return!0;var i=r.clientX,o=r.clientY,a=n.interactiveBorder,p=n.distance,s=e.top-o>("top"===t?a+p:a),l=o-e.bottom>("bottom"===t?a+p:a),c=e.left-i>("left"===t?a+p:a),f=i-e.right>("right"===t?a+p:a);return s||l||c||f}(Z(et.popper),et.popper.getBoundingClientRect(),t,et.props)&&(nt(),kt())}function ht(t){if(!gt(t))return et.props.interactive?(document.body.addEventListener("mouseleave",kt),void document.addEventListener("mousemove",H)):void kt()}function bt(t){t.target===et.reference&&(et.props.interactive&&t.relatedTarget&&et.popper.contains(t.relatedTarget)||kt())}function vt(t){h(t.target,et.props.target)&&xt(t)}function yt(t){h(t.target,et.props.target)&&kt()}function gt(t){var e="ontouchstart"in window,r=X(t.type,"touch"),n=et.props.touchHold;return e&&tt&&n&&!r||tt&&!n&&r}function wt(){var t=et.props.popperOptions,e=et.popperChildren,r=e.tooltip,o=e.arrow,a=D(t,"preventOverflow");function p(t){et.props.flip&&!et.props.flipOnUpdate&&(t.flipped&&(et.popperInstance.options.placement=t.placement),F(et.popperInstance.modifiers,!1)),r.setAttribute(g,t.placement),!1!==t.attributes[w]?r.setAttribute(w,""):r.removeAttribute(w),O&&O!==t.placement&&T&&(r.style.transition="none",requestAnimationFrame(function(){r.style.transition=""})),O=t.placement,T=et.state.isVisible;var e=Z(et.popper),n=r.style;n.top=n.bottom=n.left=n.right="",n[e]=-(et.props.distance-10)+"px";var o=a&&void 0!==a.padding?a.padding:y,p="number"==typeof o,s=i({top:p?o:o.top,bottom:p?o:o.bottom,left:p?o:o.left,right:p?o:o.right},!p&&o);s[e]=p?o+et.props.distance:(o[e]||0)+et.props.distance,et.popperInstance.modifiers.filter(function(t){return"preventOverflow"===t.name})[0].padding=s}var s=i({placement:et.props.placement},t,{modifiers:i({},t?t.modifiers:{},{preventOverflow:i({boundariesElement:et.props.boundary,padding:y},a),arrow:i({element:o,enabled:!!o},D(t,"arrow")),flip:i({enabled:et.props.flip,padding:et.props.distance+y,behavior:et.props.flipBehavior},D(t,"flip")),offset:i({offset:et.props.offset},D(t,"offset"))}),onCreate:function(e){p(e),t&&t.onCreate&&t.onCreate(e)},onUpdate:function(e){p(e),t&&t.onUpdate&&t.onUpdate(e)}});et.popperInstance=new n.a(et.reference,et.popper,s)}function xt(t){if(Et(),!et.state.isVisible){if(et.props.target)return function(t){if(t){var r=h(t.target,et.props.target);r&&!r._tippy&&(pt(r,i({},et.props,{content:N(e.content,[r]),appendTo:e.appendTo,target:"",showOnInit:!0})),xt(t))}}(t);if(A=!0,et.props.wait)return et.props.wait(et,t);ot()&&!et.state.isMounted&&document.addEventListener("mousemove",dt);var r=I(et.props.delay,0,c.delay);r?a=setTimeout(function(){Lt()},r):Lt()}}function kt(){if(Et(),!et.state.isVisible)return rt();A=!1;var t=I(et.props.delay,1,c.delay);t?s=setTimeout(function(){et.state.isVisible&&Ct()},t):l=requestAnimationFrame(function(){Ct()})}function Et(){clearTimeout(a),clearTimeout(s),cancelAnimationFrame(l)}function Ot(t){_(t=t||{},c);var e=et.props,r=z(et.reference,i({},et.props,t,{ignoreAttributes:!0}));r.ignoreAttributes=Y(t,"ignoreAttributes")?t.ignoreAttributes||!1:e.ignoreAttributes,et.props=r,(Y(t,"trigger")||Y(t,"touchHold"))&&(ft(),ct()),Y(t,"interactiveDebounce")&&(nt(),H=S(mt,t.interactiveDebounce||0)),function(t,e,r){var n=V(t),i=n.tooltip,o=n.content,a=n.backdrop,p=n.arrow;t.style.zIndex=""+r.zIndex,i.setAttribute("data-size",r.size),i.setAttribute("data-animation",r.animation),i.style.maxWidth=r.maxWidth+("number"==typeof r.maxWidth?"px":""),r.role?t.setAttribute("role",r.role):t.removeAttribute("role"),e.content!==r.content&&j(o,r),!e.animateFill&&r.animateFill?(i.appendChild(K()),i.setAttribute("data-animatefill","")):e.animateFill&&!r.animateFill&&(i.removeChild(a),i.removeAttribute("data-animatefill")),!e.arrow&&r.arrow?i.appendChild(q(r.arrowType)):e.arrow&&!r.arrow&&i.removeChild(p),e.arrow&&r.arrow&&e.arrowType!==r.arrowType&&i.replaceChild(q(r.arrowType),p),!e.interactive&&r.interactive?G(t,i):e.interactive&&!r.interactive&&function(t,e){t.removeAttribute("tabindex"),e.removeAttribute("data-interactive")}(t,i),!e.inertia&&r.inertia?R(i):e.inertia&&!r.inertia&&function(t){t.removeAttribute("data-inertia")}(i),e.theme!==r.theme&&(Q(i,"remove",e.theme),Q(i,"add",r.theme))}(et.popper,e,r),et.popperChildren=V(et.popper),et.popperInstance&&(et.popperInstance.update(),f.some(function(r){return Y(t,r)&&t[r]!==e[r]})&&(et.popperInstance.destroy(),wt(),et.state.isVisible||et.popperInstance.disableEventListeners(),et.props.followCursor&&o&&dt(o)))}function Lt(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:I(et.props.duration,0,c.duration[1]);if(!et.state.isDestroyed&&et.state.isEnabled&&(!tt||et.props.touch)&&!et.reference.hasAttribute("disabled")&&!1!==et.props.onShow(et)){et.popper.style.visibility="visible",et.state.isVisible=!0,et.props.interactive&&et.reference.classList.add(k);var e=it();W(e.concat(et.popper),0),function(t){var e=!(ot()||"initial"===et.props.followCursor&&tt);et.popperInstance?(ot()||(et.popperInstance.scheduleUpdate(),e&&et.popperInstance.enableEventListeners()),F(et.popperInstance.modifiers,et.props.flip)):(wt(),e||et.popperInstance.disableEventListeners()),et.popperInstance.reference=et.reference;var r=et.popperChildren.arrow;ot()?(r&&(r.style.margin="0"),o&&dt(o)):r&&(r.style.margin=""),tt&&o&&"initial"===et.props.followCursor&&(dt(o),r&&(r.style.margin="0")),function(t,e){var r=t.popper,n=t.options,i=n.onCreate,o=n.onUpdate;n.onCreate=n.onUpdate=function(t){!function(t){t.offsetHeight}(r),e(),o&&o(t),n.onCreate=i,n.onUpdate=o}}(et.popperInstance,t);var n=et.props.appendTo;(x="parent"===n?et.reference.parentNode:N(n,[et.reference])).contains(et.popper)||(x.appendChild(et.popper),et.props.onMount(et),et.state.isMounted=!0)}(function(){et.state.isVisible&&(ot()||et.popperInstance.update(),et.popperChildren.backdrop&&(et.popperChildren.content.style.transitionDelay=Math.round(t/12)+"ms"),et.props.sticky&&(W([et.popper],p?0:et.props.updateDuration),function t(){et.popperInstance&&et.popperInstance.scheduleUpdate(),et.state.isMounted?requestAnimationFrame(t):W([et.popper],0)}()),W([et.popper],d.updateDuration),W(e,t),B(e,"visible"),function(t,e){st(t,e)}(t,function(){et.props.aria&&et.reference.setAttribute("aria-".concat(et.props.aria),et.popper.id),et.props.onShown(et),et.state.isShown=!0}))})}}function Ct(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:I(et.props.duration,1,c.duration[1]);if(!et.state.isDestroyed&&et.state.isEnabled&&!1!==et.props.onHide(et)){et.popper.style.visibility="hidden",et.state.isVisible=!1,et.state.isShown=!1,T=!1,et.props.interactive&&et.reference.classList.remove(k);var e=it();W(e,t),B(e,"hidden"),function(t,e){st(t,function(){!et.state.isVisible&&x&&x.contains(et.popper)&&e()})}(t,function(){A||rt(),et.props.aria&&et.reference.removeAttribute("aria-".concat(et.props.aria)),et.popperInstance.disableEventListeners(),et.popperInstance.options.placement=et.props.placement,x.removeChild(et.popper),et.props.onHidden(et),et.state.isMounted=!1})}}}var st=!1;function lt(t,e){_(e||{},c),st||(document.addEventListener("click",it,!0),document.addEventListener("touchstart",et,v),window.addEventListener("blur",ot),st=!0);var r,n=i({},c,e);r=t,"[object Object]"!=={}.toString.call(r)||r.addEventListener||function(t){var e={isVirtual:!0,attributes:t.attributes||{},setAttribute:function(e,r){t.attributes[e]=r},getAttribute:function(e){return t.attributes[e]},removeAttribute:function(e){delete t.attributes[e]},hasAttribute:function(e){return e in t.attributes},addEventListener:function(){},removeEventListener:function(){},classList:{classNames:{},add:function(e){t.classList.classNames[e]=!0},remove:function(e){delete t.classList.classNames[e]},contains:function(e){return e in t.classList.classNames}}};for(var r in e)t[r]=e[r]}(t);var o=function(t){if(H(t))return[t];if(t instanceof NodeList)return m(t);if(Array.isArray(t))return t;try{return m(document.querySelectorAll(t))}catch(t){return[]}}(t).reduce(function(t,e){var r=e&&pt(e,n);return r&&t.push(r),t},[]);return H(t)?o[0]:o}lt.version="4.2.1",lt.defaults=c,lt.setDefaults=function(t){Object.keys(t).forEach(function(e){c[e]=t[e]})},lt.hideAll=$,lt.group=function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=e.delay,n=void 0===r?t[0].props.delay:r,o=e.duration,a=void 0===o?0:o;if(!t.some(function(t){return Y(t,"_originalProps")})){var p=!1;t.forEach(function(t){t._originalProps=i({},t.props)}),d()}function s(t){p=t,d()}function l(e){e._originalProps.onShow(e),t.forEach(function(t){t.set({duration:a}),t.hide()}),s(!0)}function c(t){t._originalProps.onHide(t),s(!1)}function f(t){t._originalProps.onShown(t),t.set({duration:t._originalProps.duration})}function d(){t.forEach(function(t){t.set({onShow:l,onShown:f,onHide:c,delay:p?[0,Array.isArray(n)?n[1]:n]:n,duration:p?a:t._originalProps.duration})})}},o&&setTimeout(function(){m(document.querySelectorAll("[data-tippy]")).forEach(function(t){var e=t.getAttribute("data-tippy");e&&lt(t,{content:e})})}),function(t){if(o){var e=document.createElement("style");e.type="text/css",e.textContent=t;var r=document.head,n=r.firstChild;n?r.insertBefore(e,n):r.appendChild(e)}}('.tippy-iOS{cursor:pointer!important;-webkit-tap-highlight-color:transparent}.tippy-popper{transition-timing-function:cubic-bezier(.165,.84,.44,1);max-width:calc(100% - 8px);pointer-events:none;outline:0}.tippy-popper[x-placement^=top] .tippy-backdrop{border-radius:40% 40% 0 0}.tippy-popper[x-placement^=top] .tippy-roundarrow{bottom:-7px;bottom:-6.5px;-webkit-transform-origin:50% 0;transform-origin:50% 0;margin:0 3px}.tippy-popper[x-placement^=top] .tippy-roundarrow svg{position:absolute;left:0;-webkit-transform:rotate(180deg);transform:rotate(180deg)}.tippy-popper[x-placement^=top] .tippy-arrow{border-top:8px solid #333;border-right:8px solid transparent;border-left:8px solid transparent;bottom:-7px;margin:0 3px;-webkit-transform-origin:50% 0;transform-origin:50% 0}.tippy-popper[x-placement^=top] .tippy-backdrop{-webkit-transform-origin:0 25%;transform-origin:0 25%}.tippy-popper[x-placement^=top] .tippy-backdrop[data-state=visible]{-webkit-transform:scale(1) translate(-50%,-55%);transform:scale(1) translate(-50%,-55%)}.tippy-popper[x-placement^=top] .tippy-backdrop[data-state=hidden]{-webkit-transform:scale(.2) translate(-50%,-45%);transform:scale(.2) translate(-50%,-45%);opacity:0}.tippy-popper[x-placement^=top] [data-animation=shift-toward][data-state=visible]{-webkit-transform:translateY(-10px);transform:translateY(-10px)}.tippy-popper[x-placement^=top] [data-animation=shift-toward][data-state=hidden]{opacity:0;-webkit-transform:translateY(-20px);transform:translateY(-20px)}.tippy-popper[x-placement^=top] [data-animation=perspective]{-webkit-transform-origin:bottom;transform-origin:bottom}.tippy-popper[x-placement^=top] [data-animation=perspective][data-state=visible]{-webkit-transform:perspective(700px) translateY(-10px) rotateX(0);transform:perspective(700px) translateY(-10px) rotateX(0)}.tippy-popper[x-placement^=top] [data-animation=perspective][data-state=hidden]{opacity:0;-webkit-transform:perspective(700px) translateY(0) rotateX(60deg);transform:perspective(700px) translateY(0) rotateX(60deg)}.tippy-popper[x-placement^=top] [data-animation=fade][data-state=visible]{-webkit-transform:translateY(-10px);transform:translateY(-10px)}.tippy-popper[x-placement^=top] [data-animation=fade][data-state=hidden]{opacity:0;-webkit-transform:translateY(-10px);transform:translateY(-10px)}.tippy-popper[x-placement^=top] [data-animation=shift-away][data-state=visible]{-webkit-transform:translateY(-10px);transform:translateY(-10px)}.tippy-popper[x-placement^=top] [data-animation=shift-away][data-state=hidden]{opacity:0;-webkit-transform:translateY(0);transform:translateY(0)}.tippy-popper[x-placement^=top] [data-animation=scale]{-webkit-transform-origin:bottom;transform-origin:bottom}.tippy-popper[x-placement^=top] [data-animation=scale][data-state=visible]{-webkit-transform:translateY(-10px) scale(1);transform:translateY(-10px) scale(1)}.tippy-popper[x-placement^=top] [data-animation=scale][data-state=hidden]{opacity:0;-webkit-transform:translateY(-10px) scale(.5);transform:translateY(-10px) scale(.5)}.tippy-popper[x-placement^=bottom] .tippy-backdrop{border-radius:0 0 30% 30%}.tippy-popper[x-placement^=bottom] .tippy-roundarrow{top:-7px;-webkit-transform-origin:50% 100%;transform-origin:50% 100%;margin:0 3px}.tippy-popper[x-placement^=bottom] .tippy-roundarrow svg{position:absolute;left:0;-webkit-transform:rotate(0);transform:rotate(0)}.tippy-popper[x-placement^=bottom] .tippy-arrow{border-bottom:8px solid #333;border-right:8px solid transparent;border-left:8px solid transparent;top:-7px;margin:0 3px;-webkit-transform-origin:50% 100%;transform-origin:50% 100%}.tippy-popper[x-placement^=bottom] .tippy-backdrop{-webkit-transform-origin:0 -50%;transform-origin:0 -50%}.tippy-popper[x-placement^=bottom] .tippy-backdrop[data-state=visible]{-webkit-transform:scale(1) translate(-50%,-45%);transform:scale(1) translate(-50%,-45%)}.tippy-popper[x-placement^=bottom] .tippy-backdrop[data-state=hidden]{-webkit-transform:scale(.2) translate(-50%);transform:scale(.2) translate(-50%);opacity:0}.tippy-popper[x-placement^=bottom] [data-animation=shift-toward][data-state=visible]{-webkit-transform:translateY(10px);transform:translateY(10px)}.tippy-popper[x-placement^=bottom] [data-animation=shift-toward][data-state=hidden]{opacity:0;-webkit-transform:translateY(20px);transform:translateY(20px)}.tippy-popper[x-placement^=bottom] [data-animation=perspective]{-webkit-transform-origin:top;transform-origin:top}.tippy-popper[x-placement^=bottom] [data-animation=perspective][data-state=visible]{-webkit-transform:perspective(700px) translateY(10px) rotateX(0);transform:perspective(700px) translateY(10px) rotateX(0)}.tippy-popper[x-placement^=bottom] [data-animation=perspective][data-state=hidden]{opacity:0;-webkit-transform:perspective(700px) translateY(0) rotateX(-60deg);transform:perspective(700px) translateY(0) rotateX(-60deg)}.tippy-popper[x-placement^=bottom] [data-animation=fade][data-state=visible]{-webkit-transform:translateY(10px);transform:translateY(10px)}.tippy-popper[x-placement^=bottom] [data-animation=fade][data-state=hidden]{opacity:0;-webkit-transform:translateY(10px);transform:translateY(10px)}.tippy-popper[x-placement^=bottom] [data-animation=shift-away][data-state=visible]{-webkit-transform:translateY(10px);transform:translateY(10px)}.tippy-popper[x-placement^=bottom] [data-animation=shift-away][data-state=hidden]{opacity:0;-webkit-transform:translateY(0);transform:translateY(0)}.tippy-popper[x-placement^=bottom] [data-animation=scale]{-webkit-transform-origin:top;transform-origin:top}.tippy-popper[x-placement^=bottom] [data-animation=scale][data-state=visible]{-webkit-transform:translateY(10px) scale(1);transform:translateY(10px) scale(1)}.tippy-popper[x-placement^=bottom] [data-animation=scale][data-state=hidden]{opacity:0;-webkit-transform:translateY(10px) scale(.5);transform:translateY(10px) scale(.5)}.tippy-popper[x-placement^=left] .tippy-backdrop{border-radius:50% 0 0 50%}.tippy-popper[x-placement^=left] .tippy-roundarrow{right:-12px;-webkit-transform-origin:33.33333333% 50%;transform-origin:33.33333333% 50%;margin:3px 0}.tippy-popper[x-placement^=left] .tippy-roundarrow svg{position:absolute;left:0;-webkit-transform:rotate(90deg);transform:rotate(90deg)}.tippy-popper[x-placement^=left] .tippy-arrow{border-left:8px solid #333;border-top:8px solid transparent;border-bottom:8px solid transparent;right:-7px;margin:3px 0;-webkit-transform-origin:0 50%;transform-origin:0 50%}.tippy-popper[x-placement^=left] .tippy-backdrop{-webkit-transform-origin:50% 0;transform-origin:50% 0}.tippy-popper[x-placement^=left] .tippy-backdrop[data-state=visible]{-webkit-transform:scale(1) translate(-50%,-50%);transform:scale(1) translate(-50%,-50%)}.tippy-popper[x-placement^=left] .tippy-backdrop[data-state=hidden]{-webkit-transform:scale(.2) translate(-75%,-50%);transform:scale(.2) translate(-75%,-50%);opacity:0}.tippy-popper[x-placement^=left] [data-animation=shift-toward][data-state=visible]{-webkit-transform:translateX(-10px);transform:translateX(-10px)}.tippy-popper[x-placement^=left] [data-animation=shift-toward][data-state=hidden]{opacity:0;-webkit-transform:translateX(-20px);transform:translateX(-20px)}.tippy-popper[x-placement^=left] [data-animation=perspective]{-webkit-transform-origin:right;transform-origin:right}.tippy-popper[x-placement^=left] [data-animation=perspective][data-state=visible]{-webkit-transform:perspective(700px) translateX(-10px) rotateY(0);transform:perspective(700px) translateX(-10px) rotateY(0)}.tippy-popper[x-placement^=left] [data-animation=perspective][data-state=hidden]{opacity:0;-webkit-transform:perspective(700px) translateX(0) rotateY(-60deg);transform:perspective(700px) translateX(0) rotateY(-60deg)}.tippy-popper[x-placement^=left] [data-animation=fade][data-state=visible]{-webkit-transform:translateX(-10px);transform:translateX(-10px)}.tippy-popper[x-placement^=left] [data-animation=fade][data-state=hidden]{opacity:0;-webkit-transform:translateX(-10px);transform:translateX(-10px)}.tippy-popper[x-placement^=left] [data-animation=shift-away][data-state=visible]{-webkit-transform:translateX(-10px);transform:translateX(-10px)}.tippy-popper[x-placement^=left] [data-animation=shift-away][data-state=hidden]{opacity:0;-webkit-transform:translateX(0);transform:translateX(0)}.tippy-popper[x-placement^=left] [data-animation=scale]{-webkit-transform-origin:right;transform-origin:right}.tippy-popper[x-placement^=left] [data-animation=scale][data-state=visible]{-webkit-transform:translateX(-10px) scale(1);transform:translateX(-10px) scale(1)}.tippy-popper[x-placement^=left] [data-animation=scale][data-state=hidden]{opacity:0;-webkit-transform:translateX(-10px) scale(.5);transform:translateX(-10px) scale(.5)}.tippy-popper[x-placement^=right] .tippy-backdrop{border-radius:0 50% 50% 0}.tippy-popper[x-placement^=right] .tippy-roundarrow{left:-12px;-webkit-transform-origin:66.66666666% 50%;transform-origin:66.66666666% 50%;margin:3px 0}.tippy-popper[x-placement^=right] .tippy-roundarrow svg{position:absolute;left:0;-webkit-transform:rotate(-90deg);transform:rotate(-90deg)}.tippy-popper[x-placement^=right] .tippy-arrow{border-right:8px solid #333;border-top:8px solid transparent;border-bottom:8px solid transparent;left:-7px;margin:3px 0;-webkit-transform-origin:100% 50%;transform-origin:100% 50%}.tippy-popper[x-placement^=right] .tippy-backdrop{-webkit-transform-origin:-50% 0;transform-origin:-50% 0}.tippy-popper[x-placement^=right] .tippy-backdrop[data-state=visible]{-webkit-transform:scale(1) translate(-50%,-50%);transform:scale(1) translate(-50%,-50%)}.tippy-popper[x-placement^=right] .tippy-backdrop[data-state=hidden]{-webkit-transform:scale(.2) translate(-25%,-50%);transform:scale(.2) translate(-25%,-50%);opacity:0}.tippy-popper[x-placement^=right] [data-animation=shift-toward][data-state=visible]{-webkit-transform:translateX(10px);transform:translateX(10px)}.tippy-popper[x-placement^=right] [data-animation=shift-toward][data-state=hidden]{opacity:0;-webkit-transform:translateX(20px);transform:translateX(20px)}.tippy-popper[x-placement^=right] [data-animation=perspective]{-webkit-transform-origin:left;transform-origin:left}.tippy-popper[x-placement^=right] [data-animation=perspective][data-state=visible]{-webkit-transform:perspective(700px) translateX(10px) rotateY(0);transform:perspective(700px) translateX(10px) rotateY(0)}.tippy-popper[x-placement^=right] [data-animation=perspective][data-state=hidden]{opacity:0;-webkit-transform:perspective(700px) translateX(0) rotateY(60deg);transform:perspective(700px) translateX(0) rotateY(60deg)}.tippy-popper[x-placement^=right] [data-animation=fade][data-state=visible]{-webkit-transform:translateX(10px);transform:translateX(10px)}.tippy-popper[x-placement^=right] [data-animation=fade][data-state=hidden]{opacity:0;-webkit-transform:translateX(10px);transform:translateX(10px)}.tippy-popper[x-placement^=right] [data-animation=shift-away][data-state=visible]{-webkit-transform:translateX(10px);transform:translateX(10px)}.tippy-popper[x-placement^=right] [data-animation=shift-away][data-state=hidden]{opacity:0;-webkit-transform:translateX(0);transform:translateX(0)}.tippy-popper[x-placement^=right] [data-animation=scale]{-webkit-transform-origin:left;transform-origin:left}.tippy-popper[x-placement^=right] [data-animation=scale][data-state=visible]{-webkit-transform:translateX(10px) scale(1);transform:translateX(10px) scale(1)}.tippy-popper[x-placement^=right] [data-animation=scale][data-state=hidden]{opacity:0;-webkit-transform:translateX(10px) scale(.5);transform:translateX(10px) scale(.5)}.tippy-tooltip{position:relative;color:#fff;border-radius:.25rem;font-size:.875rem;padding:.3125rem .5625rem;line-height:1.4;text-align:center;background-color:#333}.tippy-tooltip[data-size=small]{padding:.1875rem .375rem;font-size:.75rem}.tippy-tooltip[data-size=large]{padding:.375rem .75rem;font-size:1rem}.tippy-tooltip[data-animatefill]{overflow:hidden;background-color:transparent}.tippy-tooltip[data-interactive],.tippy-tooltip[data-interactive] path{pointer-events:auto}.tippy-tooltip[data-inertia][data-state=visible]{transition-timing-function:cubic-bezier(.54,1.5,.38,1.11)}.tippy-tooltip[data-inertia][data-state=hidden]{transition-timing-function:ease}.tippy-arrow,.tippy-roundarrow{position:absolute;width:0;height:0}.tippy-roundarrow{width:18px;height:7px;fill:#333;pointer-events:none}.tippy-backdrop{position:absolute;background-color:#333;border-radius:50%;width:calc(110% + 2rem);left:50%;top:50%;z-index:-1;transition:all cubic-bezier(.46,.1,.52,.98);-webkit-backface-visibility:hidden;backface-visibility:hidden}.tippy-backdrop:after{content:"";float:left;padding-top:100%}.tippy-backdrop+.tippy-content{transition-property:opacity;will-change:opacity}.tippy-backdrop+.tippy-content[data-state=visible]{opacity:1}.tippy-backdrop+.tippy-content[data-state=hidden]{opacity:0}'),e.default=lt},Zgw8:function(t,e,r){"use strict";(function(t){for(var r="undefined"!=typeof window&&"undefined"!=typeof document,n=["Edge","Trident","Firefox"],i=0,o=0;o<n.length;o+=1)if(r&&navigator.userAgent.indexOf(n[o])>=0){i=1;break}var a=r&&window.Promise?function(t){var e=!1;return function(){e||(e=!0,window.Promise.resolve().then(function(){e=!1,t()}))}}:function(t){var e=!1;return function(){e||(e=!0,setTimeout(function(){e=!1,t()},i))}};function p(t){return t&&"[object Function]"==={}.toString.call(t)}function s(t,e){if(1!==t.nodeType)return[];var r=t.ownerDocument.defaultView.getComputedStyle(t,null);return e?r[e]:r}function l(t){return"HTML"===t.nodeName?t:t.parentNode||t.host}function c(t){if(!t)return document.body;switch(t.nodeName){case"HTML":case"BODY":return t.ownerDocument.body;case"#document":return t.body}var e=s(t),r=e.overflow,n=e.overflowX,i=e.overflowY;return/(auto|scroll|overlay)/.test(r+i+n)?t:c(l(t))}var f=r&&!(!window.MSInputMethodContext||!document.documentMode),d=r&&/MSIE 10/.test(navigator.userAgent);function u(t){return 11===t?f:10===t?d:f||d}function m(t){if(!t)return document.documentElement;for(var e=u(10)?document.body:null,r=t.offsetParent||null;r===e&&t.nextElementSibling;)r=(t=t.nextElementSibling).offsetParent;var n=r&&r.nodeName;return n&&"BODY"!==n&&"HTML"!==n?-1!==["TH","TD","TABLE"].indexOf(r.nodeName)&&"static"===s(r,"position")?m(r):r:t?t.ownerDocument.documentElement:document.documentElement}function h(t){return null!==t.parentNode?h(t.parentNode):t}function b(t,e){if(!(t&&t.nodeType&&e&&e.nodeType))return document.documentElement;var r=t.compareDocumentPosition(e)&Node.DOCUMENT_POSITION_FOLLOWING,n=r?t:e,i=r?e:t,o=document.createRange();o.setStart(n,0),o.setEnd(i,0);var a,p,s=o.commonAncestorContainer;if(t!==s&&e!==s||n.contains(i))return"BODY"===(p=(a=s).nodeName)||"HTML"!==p&&m(a.firstElementChild)!==a?m(s):s;var l=h(t);return l.host?b(l.host,e):b(t,h(e).host)}function v(t){var e="top"===(arguments.length>1&&void 0!==arguments[1]?arguments[1]:"top")?"scrollTop":"scrollLeft",r=t.nodeName;if("BODY"===r||"HTML"===r){var n=t.ownerDocument.documentElement;return(t.ownerDocument.scrollingElement||n)[e]}return t[e]}function y(t,e){var r="x"===e?"Left":"Top",n="Left"===r?"Right":"Bottom";return parseFloat(t["border"+r+"Width"],10)+parseFloat(t["border"+n+"Width"],10)}function g(t,e,r,n){return Math.max(e["offset"+t],e["scroll"+t],r["client"+t],r["offset"+t],r["scroll"+t],u(10)?parseInt(r["offset"+t])+parseInt(n["margin"+("Height"===t?"Top":"Left")])+parseInt(n["margin"+("Height"===t?"Bottom":"Right")]):0)}function w(t){var e=t.body,r=t.documentElement,n=u(10)&&getComputedStyle(r);return{height:g("Height",e,r,n),width:g("Width",e,r,n)}}var x=function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")},k=function(){function t(t,e){for(var r=0;r<e.length;r++){var n=e[r];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(t,n.key,n)}}return function(e,r,n){return r&&t(e.prototype,r),n&&t(e,n),e}}(),E=function(t,e,r){return e in t?Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}):t[e]=r,t},O=Object.assign||function(t){for(var e=1;e<arguments.length;e++){var r=arguments[e];for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&(t[n]=r[n])}return t};function L(t){return O({},t,{right:t.left+t.width,bottom:t.top+t.height})}function C(t){var e={};try{if(u(10)){e=t.getBoundingClientRect();var r=v(t,"top"),n=v(t,"left");e.top+=r,e.left+=n,e.bottom+=r,e.right+=n}else e=t.getBoundingClientRect()}catch(t){}var i={left:e.left,top:e.top,width:e.right-e.left,height:e.bottom-e.top},o="HTML"===t.nodeName?w(t.ownerDocument):{},a=o.width||t.clientWidth||i.right-i.left,p=o.height||t.clientHeight||i.bottom-i.top,l=t.offsetWidth-a,c=t.offsetHeight-p;if(l||c){var f=s(t);l-=y(f,"x"),c-=y(f,"y"),i.width-=l,i.height-=c}return L(i)}function A(t,e){var r=arguments.length>2&&void 0!==arguments[2]&&arguments[2],n=u(10),i="HTML"===e.nodeName,o=C(t),a=C(e),p=c(t),l=s(e),f=parseFloat(l.borderTopWidth,10),d=parseFloat(l.borderLeftWidth,10);r&&i&&(a.top=Math.max(a.top,0),a.left=Math.max(a.left,0));var m=L({top:o.top-a.top-f,left:o.left-a.left-d,width:o.width,height:o.height});if(m.marginTop=0,m.marginLeft=0,!n&&i){var h=parseFloat(l.marginTop,10),b=parseFloat(l.marginLeft,10);m.top-=f-h,m.bottom-=f-h,m.left-=d-b,m.right-=d-b,m.marginTop=h,m.marginLeft=b}return(n&&!r?e.contains(p):e===p&&"BODY"!==p.nodeName)&&(m=function(t,e){var r=arguments.length>2&&void 0!==arguments[2]&&arguments[2],n=v(e,"top"),i=v(e,"left"),o=r?-1:1;return t.top+=n*o,t.bottom+=n*o,t.left+=i*o,t.right+=i*o,t}(m,e)),m}function T(t){if(!t||!t.parentElement||u())return document.documentElement;for(var e=t.parentElement;e&&"none"===s(e,"transform");)e=e.parentElement;return e||document.documentElement}function M(t,e,r,n){var i=arguments.length>4&&void 0!==arguments[4]&&arguments[4],o={top:0,left:0},a=i?T(t):b(t,e);if("viewport"===n)o=function(t){var e=arguments.length>1&&void 0!==arguments[1]&&arguments[1],r=t.ownerDocument.documentElement,n=A(t,r),i=Math.max(r.clientWidth,window.innerWidth||0),o=Math.max(r.clientHeight,window.innerHeight||0),a=e?0:v(r),p=e?0:v(r,"left");return L({top:a-n.top+n.marginTop,left:p-n.left+n.marginLeft,width:i,height:o})}(a,i);else{var p=void 0;"scrollParent"===n?"BODY"===(p=c(l(e))).nodeName&&(p=t.ownerDocument.documentElement):p="window"===n?t.ownerDocument.documentElement:n;var f=A(p,a,i);if("HTML"!==p.nodeName||function t(e){var r=e.nodeName;if("BODY"===r||"HTML"===r)return!1;if("fixed"===s(e,"position"))return!0;var n=l(e);return!!n&&t(n)}(a))o=f;else{var d=w(t.ownerDocument),u=d.height,m=d.width;o.top+=f.top-f.marginTop,o.bottom=u+f.top,o.left+=f.left-f.marginLeft,o.right=m+f.left}}var h="number"==typeof(r=r||0);return o.left+=h?r:r.left||0,o.top+=h?r:r.top||0,o.right-=h?r:r.right||0,o.bottom-=h?r:r.bottom||0,o}function Y(t,e,r,n,i){var o=arguments.length>5&&void 0!==arguments[5]?arguments[5]:0;if(-1===t.indexOf("auto"))return t;var a=M(r,n,o,i),p={top:{width:a.width,height:e.top-a.top},right:{width:a.right-e.right,height:a.height},bottom:{width:a.width,height:a.bottom-e.bottom},left:{width:e.left-a.left,height:a.height}},s=Object.keys(p).map(function(t){return O({key:t},p[t],{area:(e=p[t],e.width*e.height)});var e}).sort(function(t,e){return e.area-t.area}),l=s.filter(function(t){var e=t.width,n=t.height;return e>=r.clientWidth&&n>=r.clientHeight}),c=l.length>0?l[0].key:s[0].key,f=t.split("-")[1];return c+(f?"-"+f:"")}function I(t,e,r){var n=arguments.length>3&&void 0!==arguments[3]?arguments[3]:null;return A(r,n?T(e):b(e,r),n)}function S(t){var e=t.ownerDocument.defaultView.getComputedStyle(t),r=parseFloat(e.marginTop||0)+parseFloat(e.marginBottom||0),n=parseFloat(e.marginLeft||0)+parseFloat(e.marginRight||0);return{width:t.offsetWidth+n,height:t.offsetHeight+r}}function D(t){var e={left:"right",right:"left",bottom:"top",top:"bottom"};return t.replace(/left|right|bottom|top/g,function(t){return e[t]})}function X(t,e,r){r=r.split("-")[0];var n=S(t),i={width:n.width,height:n.height},o=-1!==["right","left"].indexOf(r),a=o?"top":"left",p=o?"left":"top",s=o?"height":"width",l=o?"width":"height";return i[a]=e[a]+e[s]/2-n[s]/2,i[p]=r===p?e[p]-n[l]:e[D(p)],i}function H(t,e){return Array.prototype.find?t.find(e):t.filter(e)[0]}function N(t,e,r){return(void 0===r?t:t.slice(0,function(t,e,r){if(Array.prototype.findIndex)return t.findIndex(function(t){return t[e]===r});var n=H(t,function(t){return t[e]===r});return t.indexOf(n)}(t,"name",r))).forEach(function(t){t.function&&console.warn("`modifier.function` is deprecated, use `modifier.fn`!");var r=t.function||t.fn;t.enabled&&p(r)&&(e.offsets.popper=L(e.offsets.popper),e.offsets.reference=L(e.offsets.reference),e=r(e,t))}),e}function F(t,e){return t.some(function(t){var r=t.name;return t.enabled&&r===e})}function P(t){for(var e=[!1,"ms","Webkit","Moz","O"],r=t.charAt(0).toUpperCase()+t.slice(1),n=0;n<e.length;n++){var i=e[n],o=i?""+i+r:t;if(void 0!==document.body.style[o])return o}return null}function W(t){var e=t.ownerDocument;return e?e.defaultView:window}function B(t,e,r,n){r.updateBound=n,W(t).addEventListener("resize",r.updateBound,{passive:!0});var i=c(t);return function t(e,r,n,i){var o="BODY"===e.nodeName,a=o?e.ownerDocument.defaultView:e;a.addEventListener(r,n,{passive:!0}),o||t(c(a.parentNode),r,n,i),i.push(a)}(i,"scroll",r.updateBound,r.scrollParents),r.scrollElement=i,r.eventsEnabled=!0,r}function z(){var t,e;this.state.eventsEnabled&&(cancelAnimationFrame(this.scheduleUpdate),this.state=(t=this.reference,e=this.state,W(t).removeEventListener("resize",e.updateBound),e.scrollParents.forEach(function(t){t.removeEventListener("scroll",e.updateBound)}),e.updateBound=null,e.scrollParents=[],e.scrollElement=null,e.eventsEnabled=!1,e))}function _(t){return""!==t&&!isNaN(parseFloat(t))&&isFinite(t)}function U(t,e){Object.keys(e).forEach(function(r){var n="";-1!==["width","height","top","right","bottom","left"].indexOf(r)&&_(e[r])&&(n="px"),t.style[r]=e[r]+n})}var j=r&&/Firefox/i.test(navigator.userAgent);function V(t,e,r){var n=H(t,function(t){return t.name===e}),i=!!n&&t.some(function(t){return t.name===r&&t.enabled&&t.order<n.order});if(!i){var o="`"+e+"`",a="`"+r+"`";console.warn(a+" modifier is required by "+o+" modifier in order to work, be sure to include it before "+o+"!")}return i}var R=["auto-start","auto","auto-end","top-start","top","top-end","right-start","right","right-end","bottom-end","bottom","bottom-start","left-end","left","left-start"],q=R.slice(3);function K(t){var e=arguments.length>1&&void 0!==arguments[1]&&arguments[1],r=q.indexOf(t),n=q.slice(r+1).concat(q.slice(0,r));return e?n.reverse():n}var G={FLIP:"flip",CLOCKWISE:"clockwise",COUNTERCLOCKWISE:"counterclockwise"};function J(t,e,r,n){var i=[0,0],o=-1!==["right","left"].indexOf(n),a=t.split(/(\+|\-)/).map(function(t){return t.trim()}),p=a.indexOf(H(a,function(t){return-1!==t.search(/,|\s/)}));a[p]&&-1===a[p].indexOf(",")&&console.warn("Offsets separated by white space(s) are deprecated, use a comma (,) instead.");var s=/\s*,\s*|\s+/,l=-1!==p?[a.slice(0,p).concat([a[p].split(s)[0]]),[a[p].split(s)[1]].concat(a.slice(p+1))]:[a];return(l=l.map(function(t,n){var i=(1===n?!o:o)?"height":"width",a=!1;return t.reduce(function(t,e){return""===t[t.length-1]&&-1!==["+","-"].indexOf(e)?(t[t.length-1]=e,a=!0,t):a?(t[t.length-1]+=e,a=!1,t):t.concat(e)},[]).map(function(t){return function(t,e,r,n){var i=t.match(/((?:\-|\+)?\d*\.?\d*)(.*)/),o=+i[1],a=i[2];if(!o)return t;if(0===a.indexOf("%")){var p=void 0;switch(a){case"%p":p=r;break;case"%":case"%r":default:p=n}return L(p)[e]/100*o}if("vh"===a||"vw"===a)return("vh"===a?Math.max(document.documentElement.clientHeight,window.innerHeight||0):Math.max(document.documentElement.clientWidth,window.innerWidth||0))/100*o;return o}(t,i,e,r)})})).forEach(function(t,e){t.forEach(function(r,n){_(r)&&(i[e]+=r*("-"===t[n-1]?-1:1))})}),i}var Z={placement:"bottom",positionFixed:!1,eventsEnabled:!0,removeOnDestroy:!1,onCreate:function(){},onUpdate:function(){},modifiers:{shift:{order:100,enabled:!0,fn:function(t){var e=t.placement,r=e.split("-")[0],n=e.split("-")[1];if(n){var i=t.offsets,o=i.reference,a=i.popper,p=-1!==["bottom","top"].indexOf(r),s=p?"left":"top",l=p?"width":"height",c={start:E({},s,o[s]),end:E({},s,o[s]+o[l]-a[l])};t.offsets.popper=O({},a,c[n])}return t}},offset:{order:200,enabled:!0,fn:function(t,e){var r=e.offset,n=t.placement,i=t.offsets,o=i.popper,a=i.reference,p=n.split("-")[0],s=void 0;return s=_(+r)?[+r,0]:J(r,o,a,p),"left"===p?(o.top+=s[0],o.left-=s[1]):"right"===p?(o.top+=s[0],o.left+=s[1]):"top"===p?(o.left+=s[0],o.top-=s[1]):"bottom"===p&&(o.left+=s[0],o.top+=s[1]),t.popper=o,t},offset:0},preventOverflow:{order:300,enabled:!0,fn:function(t,e){var r=e.boundariesElement||m(t.instance.popper);t.instance.reference===r&&(r=m(r));var n=P("transform"),i=t.instance.popper.style,o=i.top,a=i.left,p=i[n];i.top="",i.left="",i[n]="";var s=M(t.instance.popper,t.instance.reference,e.padding,r,t.positionFixed);i.top=o,i.left=a,i[n]=p,e.boundaries=s;var l=e.priority,c=t.offsets.popper,f={primary:function(t){var r=c[t];return c[t]<s[t]&&!e.escapeWithReference&&(r=Math.max(c[t],s[t])),E({},t,r)},secondary:function(t){var r="right"===t?"left":"top",n=c[r];return c[t]>s[t]&&!e.escapeWithReference&&(n=Math.min(c[r],s[t]-("right"===t?c.width:c.height))),E({},r,n)}};return l.forEach(function(t){var e=-1!==["left","top"].indexOf(t)?"primary":"secondary";c=O({},c,f[e](t))}),t.offsets.popper=c,t},priority:["left","right","top","bottom"],padding:5,boundariesElement:"scrollParent"},keepTogether:{order:400,enabled:!0,fn:function(t){var e=t.offsets,r=e.popper,n=e.reference,i=t.placement.split("-")[0],o=Math.floor,a=-1!==["top","bottom"].indexOf(i),p=a?"right":"bottom",s=a?"left":"top",l=a?"width":"height";return r[p]<o(n[s])&&(t.offsets.popper[s]=o(n[s])-r[l]),r[s]>o(n[p])&&(t.offsets.popper[s]=o(n[p])),t}},arrow:{order:500,enabled:!0,fn:function(t,e){var r;if(!V(t.instance.modifiers,"arrow","keepTogether"))return t;var n=e.element;if("string"==typeof n){if(!(n=t.instance.popper.querySelector(n)))return t}else if(!t.instance.popper.contains(n))return console.warn("WARNING: `arrow.element` must be child of its popper element!"),t;var i=t.placement.split("-")[0],o=t.offsets,a=o.popper,p=o.reference,l=-1!==["left","right"].indexOf(i),c=l?"height":"width",f=l?"Top":"Left",d=f.toLowerCase(),u=l?"left":"top",m=l?"bottom":"right",h=S(n)[c];p[m]-h<a[d]&&(t.offsets.popper[d]-=a[d]-(p[m]-h)),p[d]+h>a[m]&&(t.offsets.popper[d]+=p[d]+h-a[m]),t.offsets.popper=L(t.offsets.popper);var b=p[d]+p[c]/2-h/2,v=s(t.instance.popper),y=parseFloat(v["margin"+f],10),g=parseFloat(v["border"+f+"Width"],10),w=b-t.offsets.popper[d]-y-g;return w=Math.max(Math.min(a[c]-h,w),0),t.arrowElement=n,t.offsets.arrow=(E(r={},d,Math.round(w)),E(r,u,""),r),t},element:"[x-arrow]"},flip:{order:600,enabled:!0,fn:function(t,e){if(F(t.instance.modifiers,"inner"))return t;if(t.flipped&&t.placement===t.originalPlacement)return t;var r=M(t.instance.popper,t.instance.reference,e.padding,e.boundariesElement,t.positionFixed),n=t.placement.split("-")[0],i=D(n),o=t.placement.split("-")[1]||"",a=[];switch(e.behavior){case G.FLIP:a=[n,i];break;case G.CLOCKWISE:a=K(n);break;case G.COUNTERCLOCKWISE:a=K(n,!0);break;default:a=e.behavior}return a.forEach(function(p,s){if(n!==p||a.length===s+1)return t;n=t.placement.split("-")[0],i=D(n);var l=t.offsets.popper,c=t.offsets.reference,f=Math.floor,d="left"===n&&f(l.right)>f(c.left)||"right"===n&&f(l.left)<f(c.right)||"top"===n&&f(l.bottom)>f(c.top)||"bottom"===n&&f(l.top)<f(c.bottom),u=f(l.left)<f(r.left),m=f(l.right)>f(r.right),h=f(l.top)<f(r.top),b=f(l.bottom)>f(r.bottom),v="left"===n&&u||"right"===n&&m||"top"===n&&h||"bottom"===n&&b,y=-1!==["top","bottom"].indexOf(n),g=!!e.flipVariations&&(y&&"start"===o&&u||y&&"end"===o&&m||!y&&"start"===o&&h||!y&&"end"===o&&b);(d||v||g)&&(t.flipped=!0,(d||v)&&(n=a[s+1]),g&&(o=function(t){return"end"===t?"start":"start"===t?"end":t}(o)),t.placement=n+(o?"-"+o:""),t.offsets.popper=O({},t.offsets.popper,X(t.instance.popper,t.offsets.reference,t.placement)),t=N(t.instance.modifiers,t,"flip"))}),t},behavior:"flip",padding:5,boundariesElement:"viewport"},inner:{order:700,enabled:!1,fn:function(t){var e=t.placement,r=e.split("-")[0],n=t.offsets,i=n.popper,o=n.reference,a=-1!==["left","right"].indexOf(r),p=-1===["top","left"].indexOf(r);return i[a?"left":"top"]=o[r]-(p?i[a?"width":"height"]:0),t.placement=D(e),t.offsets.popper=L(i),t}},hide:{order:800,enabled:!0,fn:function(t){if(!V(t.instance.modifiers,"hide","preventOverflow"))return t;var e=t.offsets.reference,r=H(t.instance.modifiers,function(t){return"preventOverflow"===t.name}).boundaries;if(e.bottom<r.top||e.left>r.right||e.top>r.bottom||e.right<r.left){if(!0===t.hide)return t;t.hide=!0,t.attributes["x-out-of-boundaries"]=""}else{if(!1===t.hide)return t;t.hide=!1,t.attributes["x-out-of-boundaries"]=!1}return t}},computeStyle:{order:850,enabled:!0,fn:function(t,e){var r=e.x,n=e.y,i=t.offsets.popper,o=H(t.instance.modifiers,function(t){return"applyStyle"===t.name}).gpuAcceleration;void 0!==o&&console.warn("WARNING: `gpuAcceleration` option moved to `computeStyle` modifier and will not be supported in future versions of Popper.js!");var a=void 0!==o?o:e.gpuAcceleration,p=m(t.instance.popper),s=C(p),l={position:i.position},c=function(t,e){var r=t.offsets,n=r.popper,i=r.reference,o=Math.round,a=Math.floor,p=function(t){return t},s=o(i.width),l=o(n.width),c=-1!==["left","right"].indexOf(t.placement),f=-1!==t.placement.indexOf("-"),d=e?c||f||s%2==l%2?o:a:p,u=e?o:p;return{left:d(s%2==1&&l%2==1&&!f&&e?n.left-1:n.left),top:u(n.top),bottom:u(n.bottom),right:d(n.right)}}(t,window.devicePixelRatio<2||!j),f="bottom"===r?"top":"bottom",d="right"===n?"left":"right",u=P("transform"),h=void 0,b=void 0;if(b="bottom"===f?"HTML"===p.nodeName?-p.clientHeight+c.bottom:-s.height+c.bottom:c.top,h="right"===d?"HTML"===p.nodeName?-p.clientWidth+c.right:-s.width+c.right:c.left,a&&u)l[u]="translate3d("+h+"px, "+b+"px, 0)",l[f]=0,l[d]=0,l.willChange="transform";else{var v="bottom"===f?-1:1,y="right"===d?-1:1;l[f]=b*v,l[d]=h*y,l.willChange=f+", "+d}var g={"x-placement":t.placement};return t.attributes=O({},g,t.attributes),t.styles=O({},l,t.styles),t.arrowStyles=O({},t.offsets.arrow,t.arrowStyles),t},gpuAcceleration:!0,x:"bottom",y:"right"},applyStyle:{order:900,enabled:!0,fn:function(t){var e,r;return U(t.instance.popper,t.styles),e=t.instance.popper,r=t.attributes,Object.keys(r).forEach(function(t){!1!==r[t]?e.setAttribute(t,r[t]):e.removeAttribute(t)}),t.arrowElement&&Object.keys(t.arrowStyles).length&&U(t.arrowElement,t.arrowStyles),t},onLoad:function(t,e,r,n,i){var o=I(i,e,t,r.positionFixed),a=Y(r.placement,o,e,t,r.modifiers.flip.boundariesElement,r.modifiers.flip.padding);return e.setAttribute("x-placement",a),U(e,{position:r.positionFixed?"fixed":"absolute"}),r},gpuAcceleration:void 0}}},Q=function(){function t(e,r){var n=this,i=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{};x(this,t),this.scheduleUpdate=function(){return requestAnimationFrame(n.update)},this.update=a(this.update.bind(this)),this.options=O({},t.Defaults,i),this.state={isDestroyed:!1,isCreated:!1,scrollParents:[]},this.reference=e&&e.jquery?e[0]:e,this.popper=r&&r.jquery?r[0]:r,this.options.modifiers={},Object.keys(O({},t.Defaults.modifiers,i.modifiers)).forEach(function(e){n.options.modifiers[e]=O({},t.Defaults.modifiers[e]||{},i.modifiers?i.modifiers[e]:{})}),this.modifiers=Object.keys(this.options.modifiers).map(function(t){return O({name:t},n.options.modifiers[t])}).sort(function(t,e){return t.order-e.order}),this.modifiers.forEach(function(t){t.enabled&&p(t.onLoad)&&t.onLoad(n.reference,n.popper,n.options,t,n.state)}),this.update();var o=this.options.eventsEnabled;o&&this.enableEventListeners(),this.state.eventsEnabled=o}return k(t,[{key:"update",value:function(){return function(){if(!this.state.isDestroyed){var t={instance:this,styles:{},arrowStyles:{},attributes:{},flipped:!1,offsets:{}};t.offsets.reference=I(this.state,this.popper,this.reference,this.options.positionFixed),t.placement=Y(this.options.placement,t.offsets.reference,this.popper,this.reference,this.options.modifiers.flip.boundariesElement,this.options.modifiers.flip.padding),t.originalPlacement=t.placement,t.positionFixed=this.options.positionFixed,t.offsets.popper=X(this.popper,t.offsets.reference,t.placement),t.offsets.popper.position=this.options.positionFixed?"fixed":"absolute",t=N(this.modifiers,t),this.state.isCreated?this.options.onUpdate(t):(this.state.isCreated=!0,this.options.onCreate(t))}}.call(this)}},{key:"destroy",value:function(){return function(){return this.state.isDestroyed=!0,F(this.modifiers,"applyStyle")&&(this.popper.removeAttribute("x-placement"),this.popper.style.position="",this.popper.style.top="",this.popper.style.left="",this.popper.style.right="",this.popper.style.bottom="",this.popper.style.willChange="",this.popper.style[P("transform")]=""),this.disableEventListeners(),this.options.removeOnDestroy&&this.popper.parentNode.removeChild(this.popper),this}.call(this)}},{key:"enableEventListeners",value:function(){return function(){this.state.eventsEnabled||(this.state=B(this.reference,this.options,this.state,this.scheduleUpdate))}.call(this)}},{key:"disableEventListeners",value:function(){return z.call(this)}}]),t}();Q.Utils=("undefined"!=typeof window?window:t).PopperUtils,Q.placements=R,Q.Defaults=Z,e.a=Q}).call(e,r("DuR2"))}});
-//# sourceMappingURL=bundle-tooltip.js.map?id=043f6d13541a8addd049
