+++ www/assets/js/index.js
@@ -1,2 +1,2 @@
-!function(e){var t=window.webpackJsonp;window.webpackJsonp=function(n,i,o){for(var u,a,c=0,l=[];c<n.length;c++)a=n[c],r[a]&&l.push(r[a][0]),r[a]=0;for(u in i)Object.prototype.hasOwnProperty.call(i,u)&&(e[u]=i[u]);for(t&&t(n,i,o);l.length;)l.shift()()};var n={},r={2:0};function i(t){if(n[t])return n[t].exports;var r=n[t]={i:t,l:!1,exports:{}};return e[t].call(r.exports,r,r.exports,i),r.l=!0,r.exports}i.e=function(e){var t=r[e];if(0===t)return new Promise(function(e){e()});if(t)return t[2];var n=new Promise(function(n,i){t=r[e]=[n,i]});t[2]=n;var o=document.getElementsByTagName("head")[0],u=document.createElement("script");u.type="text/javascript",u.charset="utf-8",u.async=!0,u.timeout=12e4,i.nc&&u.setAttribute("nonce",i.nc),u.src=i.p+"js/bundle-"+({0:"search",1:"tooltip"}[e]||e)+".js?id="+{0:"4975eb79f9183cdae52c",1:"043f6d13541a8addd049"}[e];var a=setTimeout(c,12e4);function c(){u.onerror=u.onload=null,clearTimeout(a);var t=r[e];0!==t&&(t&&t[1](new Error("Loading chunk "+e+" failed.")),r[e]=void 0)}return u.onerror=u.onload=c,o.appendChild(u),n},i.m=e,i.c=n,i.d=function(e,t,n){i.o(e,t)||Object.defineProperty(e,t,{configurable:!1,enumerable:!0,get:n})},i.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return i.d(t,"a",t),t},i.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},i.p="/assets/",i.oe=function(e){throw console.error(e),e},i(i.s=0)}({"+Bt+":function(e,t){},0:function(e,t,n){n("vGYV"),n("SgXG"),n("P9yD"),n("+Bt+"),n("rhwI"),n("K4I0"),n("nFT7"),n("XgN2"),n("gpXz"),n("nHye"),n("m56c"),n("IFaM"),n("fDe5"),n("Md71"),n("Kn5K"),n("6J9z"),n("WgIl"),n("FnJk"),n("6314"),n("esmE"),n("2qlg"),n("w6Uh"),n("PprJ"),n("zlsc"),n("koDb"),n("i3Ly"),n("e0Aw"),n("vedj"),n("izEY"),e.exports=n("vg6f")},"2qlg":function(e,t){},6314:function(e,t){},"6J9z":function(e,t){},"7dJq":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});var r,i=(r=0,function(e){return new Promise(function(t){clearTimeout(r),r=setTimeout(t(),e)})});t.default=i},FnJk:function(e,t){},"Gnd+":function(e,t,n){"use strict";var r=function(){return function(e,t){if(Array.isArray(e))return e;if(Symbol.iterator in Object(e))return function(e,t){var n=[],r=!0,i=!1,o=void 0;try{for(var u,a=e[Symbol.iterator]();!(r=(u=a.next()).done)&&(n.push(u.value),!t||n.length!==t);r=!0);}catch(e){i=!0,o=e}finally{try{!r&&a.return&&a.return()}finally{if(i)throw o}}return n}(e,t);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),i=u(n("sFUq")),o=u(n("gc1H"));function u(e){return e&&e.__esModule?e:{default:e}}(0,i.default)().then(function(){Promise.all([n.e(1).then(n.bind(null,"IDmg"))]).then(function(e){!function(e){var t=(0,o.default)(function(){!function(e){var t=document.querySelectorAll("[data-tooltip]");if(t.length)for(var n=0,r=t.length;n<r;n++){var i=t[n],o='<div class="tippy-inner | text text-small -background:black">'+i.getAttribute("data-tooltip")+"</div>",u={animation:"shift-away",arrow:!0,delay:[150,300],distance:10,duration:[200,150],interactive:!0,maxWidth:"25rem",theme:"kirby"};u.content=o,e(i,u)}}(e)},250);new MutationObserver(t).observe(document.documentElement,{childList:!0,subtree:!0,attributes:!1,characterData:!1}),t()}(r(e,1)[0].default)})})},IFaM:function(e,t){},K4I0:function(e,t){},Kn5K:function(e,t){},Md71:function(e,t){},P9yD:function(e,t){},PSGk:function(e,t,n){"use strict";"kirbyConfig"in window&&window.kirbyConfig.assetsPath&&(n.p=window.kirbyConfig.assetsPath)},PprJ:function(e,t){},SgXG:function(e,t){},WgIl:function(e,t){},XgN2:function(e,t){},akH5:function(e,t,n){"use strict";var r=function(){return function(e,t){if(Array.isArray(e))return e;if(Symbol.iterator in Object(e))return function(e,t){var n=[],r=!0,i=!1,o=void 0;try{for(var u,a=e[Symbol.iterator]();!(r=(u=a.next()).done)&&(n.push(u.value),!t||n.length!==t);r=!0);}catch(e){i=!0,o=e}finally{try{!r&&a.return&&a.return()}finally{if(i)throw o}}return n}(e,t);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),i=u(n("sFUq")),o=u(n("7dJq"));function u(e){return e&&e.__esModule?e:{default:e}}var a='<svg viewBox="0 0 5 11" width="5" height="11" aria-hidden="true"><path d="M2.96153846,5.41538462 L0,9.13846154 L1.35384615,10.1538462 L4.73846154,5.92307692 C4.82307692,5.75384615 4.90769231,5.58461538 4.90769231,5.41538462 C4.90769231,5.24615385 4.82307692,5.07692308 4.73846154,4.90769231 L1.35384615,0.676923077 L0,1.69230769 L2.96153846,5.41538462 Z"/></svg>';(0,i.default)(function(){var e=document.querySelector(".js-menu-search");if(e){var t=document.querySelector(".js-menu-search-input");Promise.all([n.e(0).then(n.bind(null,"l3sP")),n.e(0).then(n.bind(null,"S6+x"))]).then(function(n){var i=r(n,2),u=i[0],c=i[1],l=u("S7OGBIAJTV","d161a2f4cd2d69247c529a3371ad3050").initIndex("getkirby-3"),s="",f=[],d=new c(t,{list:[],filter:function(){return!0},sort:!1,item:function(e){var t=document.createElement("li");return"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd"===e.label?(t.classList.add("menu-search-view-all"),t.innerHTML="<strong>View all results</strong>"+a):"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd"===e.label?(t.classList.add("menu-search-error"),t.innerHTML="<strong>Sorry, an error occured. Please try advanced search instead.</strong>"+a):t.innerHTML="<strong>"+e.label+"</strong> <small>"+e.value+"</small>",t}});t.addEventListener("focus",function(){document.documentElement.classList.add("is-menu-search-open")}),t.addEventListener("blur",function(){document.documentElement.classList.remove("is-menu-search-open")}),t.addEventListener("keypress",function(t){(t.key&&"Enter"===t.key||13===t.keyCode)&&e.submit()}),t.addEventListener("keyup",function(){var e=t.value.trim();return e===s||(""===e?(s="",f=[],d.list=f,d.evaluate(),!0):(s=e,e.length<=2||void(0,o.default)(250).then(function(){var n={hitsPerPage:5},r=t.getAttribute("data-filters");r&&(n.filters=r),l.search(e,n,function(e,t){if(f=[],e)return"console"in window&&console.error("Quicksearch error",e),f.push({label:"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd",value:0}),d.list=f,d.evaluate(),void d.open();for(var n=0,r=t.hits.length;n<r;n++){var i=t.hits[n];f.push({label:i.title,value:i.objectID})}t.nbHits>0&&t.nbHits>t.hitsPerPage&&f.push({label:"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd",value:t.nbHits}),d.list=f,d.evaluate(),d.open()})})))}),t.addEventListener("awesomplete-select",function(t){t.preventDefault(),"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd"===t.text.label||"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd"===t.text.label?e.submit():window.location.href="/"+t.text.value})})}})},e0Aw:function(e,t){},esmE:function(e,t){},fDe5:function(e,t){},gc1H:function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(e,t){var n=!1;return function(){n||(e(),n=!0,setTimeout(function(){n=!1},t))}}},gpXz:function(e,t){},i3Ly:function(e,t){},izEY:function(e,t){},koDb:function(e,t){},m56c:function(e,t){},nFT7:function(e,t){},nHye:function(e,t){},rhwI:function(e,t){},sFUq:function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:function(){};return new Promise(function(t){"loading"!==document.readyState?(e(),t()):document.addEventListener("DOMContentLoaded",function(){e(),t()})})}},vGYV:function(e,t,n){"use strict";n("PSGk"),n("akH5"),n("Gnd+")},vedj:function(e,t){},vg6f:function(e,t){},w6Uh:function(e,t){},zlsc:function(e,t){}});
+!function(e){var n=window.webpackJsonp;window.webpackJsonp=function(t,o,i){for(var u,a,c=0,s=[];c<t.length;c++)a=t[c],r[a]&&s.push(r[a][0]),r[a]=0;for(u in o)Object.prototype.hasOwnProperty.call(o,u)&&(e[u]=o[u]);for(n&&n(t,o,i);s.length;)s.shift()()};var t={},r={1:0};function o(n){if(t[n])return t[n].exports;var r=t[n]={i:n,l:!1,exports:{}};return e[n].call(r.exports,r,r.exports,o),r.l=!0,r.exports}o.e=function(e){var n=r[e];if(0===n)return new Promise(function(e){e()});if(n)return n[2];var t=new Promise(function(t,o){n=r[e]=[t,o]});n[2]=t;var i=document.getElementsByTagName("head")[0],u=document.createElement("script");u.type="text/javascript",u.charset="utf-8",u.async=!0,u.timeout=12e4,o.nc&&u.setAttribute("nonce",o.nc),u.src=o.p+"js/bundle-"+({0:"search"}[e]||e)+".js?id="+{0:"4975eb79f9183cdae52c"}[e];var a=setTimeout(c,12e4);function c(){u.onerror=u.onload=null,clearTimeout(a);var n=r[e];0!==n&&(n&&n[1](new Error("Loading chunk "+e+" failed.")),r[e]=void 0)}return u.onerror=u.onload=c,i.appendChild(u),t},o.m=e,o.c=t,o.d=function(e,n,t){o.o(e,n)||Object.defineProperty(e,n,{configurable:!1,enumerable:!0,get:t})},o.n=function(e){var n=e&&e.__esModule?function(){return e.default}:function(){return e};return o.d(n,"a",n),n},o.o=function(e,n){return Object.prototype.hasOwnProperty.call(e,n)},o.p="/assets/",o.oe=function(e){throw console.error(e),e},o(o.s=0)}({"+Bt+":function(e,n){},0:function(e,n,t){t("vGYV"),t("SgXG"),t("P9yD"),t("+Bt+"),t("rhwI"),t("K4I0"),t("nFT7"),t("XgN2"),t("gpXz"),t("nHye"),t("m56c"),t("IFaM"),t("fDe5"),t("Md71"),t("Kn5K"),t("6J9z"),t("WgIl"),t("FnJk"),t("6314"),t("esmE"),t("2qlg"),t("w6Uh"),t("PprJ"),t("zlsc"),t("koDb"),t("i3Ly"),t("e0Aw"),t("vedj"),t("izEY"),e.exports=t("vg6f")},"2qlg":function(e,n){},6314:function(e,n){},"6J9z":function(e,n){},"7dJq":function(e,n,t){"use strict";Object.defineProperty(n,"__esModule",{value:!0});var r,o=(r=0,function(e){return new Promise(function(n){clearTimeout(r),r=setTimeout(n(),e)})});n.default=o},FnJk:function(e,n){},IFaM:function(e,n){},K4I0:function(e,n){},Kn5K:function(e,n){},Md71:function(e,n){},P9yD:function(e,n){},PSGk:function(e,n,t){"use strict";"kirbyConfig"in window&&window.kirbyConfig.assetsPath&&(t.p=window.kirbyConfig.assetsPath)},PprJ:function(e,n){},SgXG:function(e,n){},WgIl:function(e,n){},XgN2:function(e,n){},akH5:function(e,n,t){"use strict";var r=function(){return function(e,n){if(Array.isArray(e))return e;if(Symbol.iterator in Object(e))return function(e,n){var t=[],r=!0,o=!1,i=void 0;try{for(var u,a=e[Symbol.iterator]();!(r=(u=a.next()).done)&&(t.push(u.value),!n||t.length!==n);r=!0);}catch(e){o=!0,i=e}finally{try{!r&&a.return&&a.return()}finally{if(o)throw i}}return t}(e,n);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),o=u(t("sFUq")),i=u(t("7dJq"));function u(e){return e&&e.__esModule?e:{default:e}}var a='<svg viewBox="0 0 5 11" width="5" height="11" aria-hidden="true"><path d="M2.96153846,5.41538462 L0,9.13846154 L1.35384615,10.1538462 L4.73846154,5.92307692 C4.82307692,5.75384615 4.90769231,5.58461538 4.90769231,5.41538462 C4.90769231,5.24615385 4.82307692,5.07692308 4.73846154,4.90769231 L1.35384615,0.676923077 L0,1.69230769 L2.96153846,5.41538462 Z"/></svg>';(0,o.default)(function(){var e=document.querySelector(".js-menu-search");if(e){var n=document.querySelector(".js-menu-search-input");Promise.all([t.e(0).then(t.bind(null,"l3sP")),t.e(0).then(t.bind(null,"S6+x"))]).then(function(t){var o=r(t,2),u=o[0],c=o[1],s=u("S7OGBIAJTV","d161a2f4cd2d69247c529a3371ad3050").initIndex("getkirby-3"),l="",f=[],d=new c(n,{list:[],filter:function(){return!0},sort:!1,item:function(e){var n=document.createElement("li");return"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd"===e.label?(n.classList.add("menu-search-view-all"),n.innerHTML="<strong>View all results</strong>"+a):"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd"===e.label?(n.classList.add("menu-search-error"),n.innerHTML="<strong>Sorry, an error occured. Please try advanced search instead.</strong>"+a):n.innerHTML="<strong>"+e.label+"</strong> <small>"+e.value+"</small>",n}});n.addEventListener("focus",function(){document.documentElement.classList.add("is-menu-search-open")}),n.addEventListener("blur",function(){document.documentElement.classList.remove("is-menu-search-open")}),n.addEventListener("keypress",function(n){(n.key&&"Enter"===n.key||13===n.keyCode)&&e.submit()}),n.addEventListener("keyup",function(){var e=n.value.trim();return e===l||(""===e?(l="",f=[],d.list=f,d.evaluate(),!0):(l=e,e.length<=2||void(0,i.default)(250).then(function(){var t={hitsPerPage:5},r=n.getAttribute("data-filters");r&&(t.filters=r),s.search(e,t,function(e,n){if(f=[],e)return"console"in window&&console.error("Quicksearch error",e),f.push({label:"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd",value:0}),d.list=f,d.evaluate(),void d.open();for(var t=0,r=n.hits.length;t<r;t++){var o=n.hits[t];f.push({label:o.title,value:o.objectID})}n.nbHits>0&&n.nbHits>n.hitsPerPage&&f.push({label:"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd",value:n.nbHits}),d.list=f,d.evaluate(),d.open()})})))}),n.addEventListener("awesomplete-select",function(n){n.preventDefault(),"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd"===n.text.label||"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd"===n.text.label?e.submit():window.location.href="/"+n.text.value})})}})},e0Aw:function(e,n){},esmE:function(e,n){},fDe5:function(e,n){},gpXz:function(e,n){},i3Ly:function(e,n){},izEY:function(e,n){},koDb:function(e,n){},m56c:function(e,n){},nFT7:function(e,n){},nHye:function(e,n){},rhwI:function(e,n){},sFUq:function(e,n,t){"use strict";Object.defineProperty(n,"__esModule",{value:!0}),n.default=function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:function(){};return new Promise(function(n){"loading"!==document.readyState?(e(),n()):document.addEventListener("DOMContentLoaded",function(){e(),n()})})}},vGYV:function(e,n,t){"use strict";t("PSGk"),t("akH5")},vedj:function(e,n){},vg6f:function(e,n){},w6Uh:function(e,n){},zlsc:function(e,n){}});
 //# sourceMappingURL=index.js.map
