+++ www/assets/mix-manifest.json
@@ -1,10 +1,8 @@
 {
     "/js/bundle-search.js": "/js/bundle-search.js?id=3dbee769c74b69538fed",
-    "/js/bundle-search.js.map": "/js/bundle-search.js.map?id=ef321cf1e70c8256089b",
-    "/js/bundle-tooltip.js": "/js/bundle-tooltip.js?id=3c2b07b07ba6c0f942ce",
-    "/js/bundle-tooltip.js.map": "/js/bundle-tooltip.js.map?id=ef0323c533cb332d13f1",
-    "/js/index.js": "/js/index.js?id=4f153610c0b22ffa855b",
-    "/css/index.css": "/css/index.css?id=a43bed9e3a02d411f336",
+    "/js/bundle-search.js.map": "/js/bundle-search.js.map?id=7058df8957feb9660198",
+    "/js/index.js": "/js/index.js?id=f37cdbad4d7087addb29",
+    "/css/index.css": "/css/index.css?id=2a1371afb911d5fd2018",
     "/css/templates/archive.css": "/css/templates/archive.css?id=d3fe33b8452c5a10da9c",
     "/css/templates/buy.css": "/css/templates/buy.css?id=cac5746063868f0329fd",
     "/css/templates/cases.css": "/css/templates/cases.css?id=d31a56a1bcb5279a2a4d",
@@ -33,8 +31,8 @@
     "/css/templates/text.css": "/css/templates/text.css?id=c265590bb71d15f06d1e",
     "/css/templates/try.css": "/css/templates/try.css?id=ddf925c7ca5e12567ed0",
     "/css/templates/why.css": "/css/templates/why.css?id=f4b2feed2d9fe592c103",
-    "/js/index.js.map": "/js/index.js.map?id=4a856a43863b911c76b7",
-    "/css/index.css.map": "/css/index.css.map?id=abf0039b5ae3d5cf7e83",
+    "/js/index.js.map": "/js/index.js.map?id=d10e4315c22fdee7bc60",
+    "/css/index.css.map": "/css/index.css.map?id=057a81695bd2baf7ffa6",
     "/css/templates/archive.css.map": "/css/templates/archive.css.map?id=6c2694ea745675954f02",
     "/css/templates/buy.css.map": "/css/templates/buy.css.map?id=a1d87ca982c004494df7",
     "/css/templates/cases.css.map": "/css/templates/cases.css.map?id=1d73cd561149712db359",
